+++ EasyTicket/easy-ticket-web/app/views/seatplan/list.blade.php
@@ -9,29 +9,9 @@
             <!-- BEGIN PAGE HEADER-->
                <div class="row-fluid">
                   <div class="span12">
-                     <!-- BEGIN STYLE CUSTOMIZER -->
-                     <div class="color-panel hidden-phone">
-                        <div class="color-mode-icons icon-color"></div>
-                        <div class="color-mode-icons icon-color-close"></div>
-                        <div class="color-mode">
-                           <p>THEME COLOR</p>
-                           <ul class="inline">
-                              <li class="color-black current color-default" data-style="default"></li>
-                              <li class="color-blue" data-style="blue"></li>
-                              <li class="color-brown" data-style="brown"></li>
-                              <li class="color-purple" data-style="purple"></li>
-                              <li class="color-white color-light" data-style="light"></li>
-                           </ul>
-                           <label class="hidden-phone">
-                           <input type="checkbox" class="header" checked value="" />
-                           <span class="color-mode-label">Fixed Header</span>
-                           </label>                   
-                        </div>
-                     </div>
-                     <!-- END BEGIN STYLE CUSTOMIZER -->    
                      <!-- BEGIN PAGE TITLE & BREADCRUMB-->        
                      <h3 class="page-title">
-                        Dashboard            
+                        Seat Plan List            
                         <small></small>
                      </h3>
                      <ul class="breadcrumb">
@@ -40,7 +20,7 @@
                            <a href="/">Home</a> 
                            <i class="icon-angle-right"></i>
                         </li>
-                        <li><a href="/dashboard">Dashboard</a></li>
+                        <li><a href="#">Seat Plan List</a></li>
                         
                      </ul>
                      <!-- END PAGE TITLE & BREADCRUMB-->
@@ -88,7 +68,9 @@
                                        <th>Ticket Type</th>
                                        <th class="span6">Operator</th>
                                        <th>Edit</th>
-                                       <th>Delete</th>
+                                       @if(Auth::user()->role!=2)
+                                          <th>Delete</th>
+                                       @endif
                                        <th>View</th>
                                     </tr>
                                  </thead>
@@ -106,9 +88,11 @@
                                             <td style="text-align:center;">
                                               <a href="/seatplan/update/{{ $seatplan['id'] }}"  class="btn green button-submit">Edit</a><br><br>
                                             </td>
-                                            <td>
-                                              <a href="deleteseatplan/{{ $seatplan['id'] }}"   class="btn green delete button-submit">Delete</a>
-                                            </td>
+                                             @if(Auth::user()->role!=2)
+                                                <td>
+                                                   <a href="deleteseatplan/{{ $seatplan['id'] }}"   class="btn green delete button-submit">Delete</a>
+                                                </td>
+                                             @endif
                                             <td>
                                                 <a href="/seatplandetail/{{$seatplan['id']}}/seat_plan_id?seatplan_id={{$seatplan['id']}}&operator_id={{$seatplan['operator_id']}}&name={{$seatplan['name']}}&ticket_type_id={{$seatplan['ticket_type_id']}}"   class="btn green button-submit">Detail</a>
                                             </td>
